* Added class that allows the transfomation of composit sub-metrics and metrics' measures

+++ trunk/src/platform/toolkit/metrics/library/core/ExponentiallyDecayingResevoir.cs
@@ -224,7 +224,7 @@ namespace Nohros.Metrics
     }
 
     long TimeInSeconds(long ticks) {
-      return ticks.ToSeconds(TimeUnit.Ticks);
+      return (long) ticks.ToSeconds(TimeUnit.Ticks);
     }
   }
 }
